сделал что-б не удалял чужое (вообще мелочь но типо мало ли?)

+++ app/Http/Controllers/ListsController.php
@@ -81,6 +81,9 @@ class ListsController extends Controller
 
     public function destroy(string $id)
     {
+        if (Auth::user()->isAdmin()==false && Lists::where('id', $id)->first()->author != Auth::id())  {
+            return 'acces denied !!!';
+        }
         Lists::find($id)->delete();
         return redirect('dashboard');
     }
